% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_gpd.R, R/updating_alcohol.R
\name{allocate_gpd}
\alias{allocate_gpd}
\title{Allocate individuals a grams per day of consumption - based on categories}
\usage{
allocate_gpd(basepop, percentreduction)

allocate_gpd(basepop, percentreduction)
}
\description{
Allocate individuals a grams per day of consumption - based on categories

Allocate individuals a grams per day of consumption - based on categories
}
\examples{
allocate_gpd
simulating reducing alcohol consumption
}
\keyword{alcohol}
\keyword{allocate}
\keyword{reduce}
