% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_gramsperday.R
\name{allocate_gramsperday}
\alias{allocate_gramsperday}
\title{Allocate individuals a grams per day of consumption - based on categories}
\usage{
allocate_gramsperday(basepop)
}
\description{
Allocate individuals a grams per day of consumption - based on categories
}
\examples{
allocate_gpd
}
\keyword{alcohol}
\keyword{allocate}
