% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_gramsperday_sampled.R
\name{allocate_gramsperday_sampled}
\alias{allocate_gramsperday_sampled}
\title{Allocate individuals a grams per day of consumption - based on categories}
\usage{
allocate_gramsperday_sampled(basepop, y, catcontmodel)
}
\description{
Allocate individuals a grams per day of consumption - based on categories
}
\examples{
allocate_gramsperday_sampled
}
\keyword{alcohol}
\keyword{allocate}
