% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_death_counts.R
\name{apply_death_counts}
\alias{apply_death_counts}
\title{Apply death counts - converted to rates in each year of simulation}
\usage{
apply_death_counts(basepop, death_counts, y, diseases)
}
\description{
Apply death counts - converted to rates in each year of simulation
}
\examples{
apply_death_counts
}
\keyword{microsimulation}
