% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_death_rates.R
\name{apply_death_rates}
\alias{apply_death_rates}
\title{Apply death rates in each year of simulation}
\usage{
apply_death_rates(basepop, death_rates, y, diseases)
}
\description{
Apply death rates in each year of simulation
}
\examples{
apply_death_rates
}
\keyword{microsimulation}
