% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_tax_policy.R
\name{apply_tax_policy}
\alias{apply_tax_policy}
\title{Apply taxation policy}
\usage{
apply_tax_policy(
  data = basepop,
  scenario = scenario,
  participation = participation,
  part_elasticity = part_elasticity,
  prob_alcohol_transitions = prob_alcohol_transitions,
  cons_elasticity = cons_elasticity,
  cons_elasticity_se = cons_elasticity_se,
  r_sim_obs = r_sim_obs
)
}
\description{
This function simulates tax policy
}
\keyword{policy}
\keyword{tax}
