% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_acute_hep.R
\name{assign_acute_hep}
\alias{assign_acute_hep}
\title{Assign new acute cases of hepatitis in each year of the simulation
based on IHME data on incidence of new hepatitis B and C infections}
\usage{
assign_acute_hep(microsim, Hep, distribution, y)
}
\description{
Assign new acute cases of hepatitis in each year of the simulation
based on IHME data on incidence of new hepatitis B and C infections
}
\examples{
assign acute hepatitis - throughout the simulation
}
\keyword{acute}
\keyword{assign}
\keyword{hepatitis}
