% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_baseline_hepatitis.R
\name{assign_baseline_hepatitis}
\alias{assign_baseline_hepatitis}
\title{Allocate individuals a chronic hepatitis B or C status - based on NHANES data}
\usage{
assign_baseline_hepatitis(basepop)
}
\description{
Allocate individuals a chronic hepatitis B or C status - based on NHANES data
}
\examples{
assign_baseline_hepatitis
}
\keyword{assign}
\keyword{hepatitis}
