% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_beverage_preferences.R
\name{assign_beverage_preferences}
\alias{assign_beverage_preferences}
\title{Assign each individual in the microsimulation with a proportion of beer, wine, liquor, coolers}
\usage{
assign_beverage_preferences(data)
}
\description{
Assign each individual in the microsimulation with a proportion of beer, wine, liquor, coolers
}
\examples{
assign beverage preferences
}
\keyword{assign}
\keyword{beverage}
\keyword{preferences}
