% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_chronic_hep.R
\name{assign_chronic_hep}
\alias{assign_chronic_hep}
\title{Assign new chronic cases of hepatitis in each year of the simulation
based on acute hepatitis plus clearance rate}
\usage{
assign_chronic_hep(microsim)
}
\description{
Assign new chronic cases of hepatitis in each year of the simulation
based on acute hepatitis plus clearance rate
}
\examples{
assign chronic hepatitis - throughout the simulation
}
\keyword{assign}
\keyword{chronic}
\keyword{hepatitis}
