% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUD_interaction.R
\name{AUDInteraction}
\alias{AUDInteraction}
\title{Calculates RR of AUD - with SES interaction effects}
\usage{
AUDInteraction(data, lhs)
}
\description{
Calculates RR of AUD - with SES interaction effects
}
\examples{
AUD including SES interaction effects
}
\keyword{AUD}
\keyword{effects}
\keyword{interaction}
