% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_base_rate.R
\name{calculate_base_rate}
\alias{calculate_base_rate}
\title{Calculates baseline mortality rates for different diseases}
\usage{
calculate_base_rate(data, base_rates, diseases)
}
\description{
Calculates baseline mortality rates for different diseases
}
\examples{
base rates
}
\keyword{base}
\keyword{calculate}
\keyword{rate}
