% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirrhosis_heavy_use.R
\name{cirrhosis_heavy_use}
\alias{cirrhosis_heavy_use}
\title{Calculates RR of liver cirrhosis through heavy alcohol use}
\usage{
cirrhosis_heavy_use(data, lhsSample)
}
\description{
Calculates RR of liver cirrhosis through heavy alcohol use
}
\examples{
cirrhosis_heavy_use
}
\keyword{alcohol}
\keyword{cirrhosis}
\keyword{use}
