% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirrhosis_metabolic.R
\name{cirrhosis_metabolic}
\alias{cirrhosis_metabolic}
\title{Calculates RR of liver cirrhosis through metabolic pathway}
\usage{
cirrhosis_metabolic(data, lhsSample)
}
\description{
Calculates RR of liver cirrhosis through metabolic pathway
}
\examples{
cirrhosis_metabolic pathway
}
\keyword{cirrhosis}
\keyword{metabolic}
\keyword{pathway}
