% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirrhosis_all.R
\name{CirrhosisAll}
\alias{CirrhosisAll}
\title{Calculates RR of liver cirrhosis through all type pathway}
\usage{
CirrhosisAll(data, lhs)
}
\description{
Calculates RR of liver cirrhosis through all type pathway
}
\examples{
all type pathway pathway
}
\keyword{all}
\keyword{cirrhosis}
\keyword{pathway}
\keyword{type}
