% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirrhosis_all_interaction.R
\name{CirrhosisAllInteraction}
\alias{CirrhosisAllInteraction}
\title{Calculates RR of liver cirrhosis through all type pathway - with different SES effects}
\usage{
CirrhosisAllInteraction(data, lhs)
}
\description{
Calculates RR of liver cirrhosis through all type pathway - with different SES effects
}
\examples{
all type pathway pathway including SES interaction
}
\keyword{all}
\keyword{cirrhosis}
\keyword{effects}
\keyword{interaction}
\keyword{pathway}
\keyword{type}
