% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirrhosis_hepatitis.R
\name{CirrhosisHepatitis}
\alias{CirrhosisHepatitis}
\title{Calculates RR of liver cirrhosis through hepatitis pathway}
\usage{
CirrhosisHepatitis(data, lhsSample)
}
\description{
Calculates RR of liver cirrhosis through hepatitis pathway
}
\examples{
hepatitis pathway
}
\keyword{cirrhosis}
\keyword{hepatitis}
\keyword{pathway}
