% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_alcohol_categories.R
\name{code_alcohol_categories}
\alias{code_alcohol_categories}
\title{Codes alcohol categories for the model}
\usage{
code_alcohol_categories(data)
}
\description{
Codes alcohol categories for the model
}
\examples{
code alcohol categories
}
\keyword{alcohol}
\keyword{categories}
