% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_migration_counts.R
\name{compare_migration_count}
\alias{compare_migration_count}
\title{Compares population counts and adjusts migration counts for any given year}
\usage{
compare_migration_count(Output, year, migration_counts)
}
\description{
Compares population counts and adjusts migration counts for any given year
}
\examples{
compare_migration_count
}
\keyword{migration}
