% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education_setup.R
\name{education_setup}
\alias{education_setup}
\title{Set up education data for simulation}
\usage{
education_setup(basepop, y)
}
\arguments{
\item{basepop, }{y takes the simulated population in each year}
}
\description{
This function allows you to express your love of cats.
}
\examples{
education_setup()
}
\keyword{education}
