% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_initial_education.R
\name{fix_initial_education}
\alias{fix_initial_education}
\title{fix initial values of education at individual age - for education transitions calibration}
\usage{
fix_initial_education(basepop)
}
\description{
fix initial values of education at individual age - for education transitions calibration
}
\examples{
fix initial education
}
\keyword{implausibility}
\keyword{markov}
\keyword{microsimulation}
\keyword{model}
