% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_years_somecollege.R
\name{fix_years_somecollege}
\alias{fix_years_somecollege}
\title{Loads eduacation transition rate data for SIMAH or CASCADE versions}
\usage{
fix_years_somecollege(basepop)
}
\description{
Loads eduacation transition rate data for SIMAH or CASCADE versions
}
\examples{
load_education_transitions
}
\keyword{education}
\keyword{rates}
\keyword{transition}
