% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_alcoholcat_summary.R
\name{generate_alcoholcat_summary}
\alias{generate_alcoholcat_summary}
\title{summarises the data for alcohol categories}
\usage{
generate_alcoholcat_summary(data)
}
\description{
summarises the data for alcohol categories
}
\examples{
summarise alcohol cat data by grouping variables
}
\keyword{alcohol}
\keyword{categories}
