% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_weights_IPF.R
\name{generate_weights_IPF}
\alias{generate_weights_IPF}
\title{generates the IPF weights for each selected state}
\usage{
generate_weights_IPF(data, selectedstate, cons)
}
\description{
generates the IPF weights for each selected state
}
\examples{
setup_for_IPF
}
\keyword{brfss}
