% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertension.R
\name{HYPHD}
\alias{HYPHD}
\title{Calculates RR of hypertensive heart disease}
\usage{
HYPHD(data, lhs)
}
\description{
Calculates RR of hypertensive heart disease
}
\examples{
Hypertensive heart disease mortality risk
}
\keyword{HYPERTENSION}
