% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHD.R
\name{IHD}
\alias{IHD}
\title{Calculates RR of ischaemic heart disease}
\usage{
IHD(data, lhs)
}
\description{
Calculates RR of ischaemic heart disease
}
\examples{
Ischaemic heart disease mortality risk
}
\keyword{IHD}
