% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHD_interaction.R
\name{IHDInteraction}
\alias{IHDInteraction}
\title{Calculates RR of IHD - with SES interaction effects}
\usage{
IHDInteraction(data, lhs)
}
\description{
Calculates RR of IHD - with SES interaction effects
}
\examples{
IHD including SES interaction effects
}
\keyword{IHD}
\keyword{effects}
\keyword{interaction}
