% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inward_births.R
\name{inward_births}
\alias{inward_births}
\title{Adds new births- estimating rates version}
\usage{
inward_births(basepop, migration_counts, y, brfss, model)
}
\description{
Adds new births- estimating rates version
}
\examples{
inward_births_estimate_rate
}
\keyword{microsimulation}
