% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inward_births_estimate_rate.R
\name{inward_births_estimate_rate}
\alias{inward_births_estimate_rate}
\title{Adds new births- estimating rates version}
\usage{
inward_births_estimate_rate(basepop, migration_counts, y, brfss)
}
\description{
Adds new births- estimating rates version
}
\examples{
inward_births_estimate_rate
}
\keyword{microsimulation}
