% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inward_births_rate.R
\name{inward_births_rate}
\alias{inward_births_rate}
\title{Adds new births- using rates version}
\usage{
inward_births_rate(basepop, migration_rates, y, brfss, model)
}
\description{
Adds new births- using rates version
}
\examples{
inward_births_estimate_rate
}
\keyword{microsimulation}
