% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inward_migration.R
\name{inward_migration}
\alias{inward_migration}
\title{Adds new migrants in each year from existing data pool}
\usage{
inward_migration(basepop, migration_rates, y, brfss)
}
\description{
Adds new migrants in each year from existing data pool
}
\examples{
inward_migration
}
\keyword{microsimulation}
