% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istroke.R
\name{ISTR}
\alias{ISTR}
\title{Calculates RR of ischaemic stroke}
\usage{
ISTR(data, lhs)
}
\description{
Calculates RR of ischaemic stroke
}
\examples{
ISCHAEMIC STROKE mortality risk ##NOT CORRECT - function is for haemorrhagic stroke
}
\keyword{STROKE}
