% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_alcohol_transitions.R
\name{load_alcohol_transitions}
\alias{load_alcohol_transitions}
\title{Loads alcohol transition rate data for SIMAH or CASCADE versions}
\usage{
load_alcohol_transitions(basepop, brfss, WorkingDirectory)
}
\description{
Loads alcohol transition rate data for SIMAH or CASCADE versions
}
\examples{
load_alcohol_transitions
}
\keyword{alcohol}
\keyword{rates}
\keyword{transition}
