% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_brfss.R
\name{load_brfss}
\alias{load_brfss}
\title{Loads BRFSS data for new migrants and 18-year-olds}
\usage{
load_brfss(model = "SIMAH", SelectedState)
}
\description{
Loads BRFSS data for new migrants and 18-year-olds
}
\examples{
load_brfss
}
\keyword{brfss}
