% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_death_counts.R
\name{load_death_counts}
\alias{load_death_counts}
\title{Loads death count data for SIMAH or CASCADE versions}
\usage{
load_death_counts(model = "SIMAH", proportion, SelectedState, DataDirectory)
}
\description{
Loads death count data for SIMAH or CASCADE versions
}
\examples{
load_death_counts
}
\keyword{counts}
\keyword{death}
