% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_death_rates.R
\name{load_death_rates}
\alias{load_death_rates}
\title{Loads death rate data for SIMAH or CASCADE versions}
\usage{
load_death_rates(model = "SIMAH", proportion, SelectedState, DataDirectory)
}
\description{
Loads death rate data for SIMAH or CASCADE versions
}
\examples{
load_death_rates
}
\keyword{death}
\keyword{rates}
