% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_education_transitions.R
\name{load_education_transitions}
\alias{load_education_transitions}
\title{Loads eduacation transition rate data for SIMAH or CASCADE versions}
\usage{
load_education_transitions(SelectedState, basepop, brfss, WorkingDirectory)
}
\description{
Loads eduacation transition rate data for SIMAH or CASCADE versions
}
\examples{
load_education_transitions
}
\keyword{education}
\keyword{rates}
\keyword{transition}
