% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hepatitis_data.R
\name{load_hepatitis_data}
\alias{load_hepatitis_data}
\title{Load hepatitis B and C incidence data - to allocate new cases of hepatitis in each year}
\usage{
load_hepatitis_data(SelectedState, proportion)
}
\description{
Load hepatitis B and C incidence data - to allocate new cases of hepatitis in each year
}
\examples{
load_hepatitis_data
}
\keyword{hepatitis}
\keyword{load}
