% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_migration_counts.R
\name{load_migration_counts}
\alias{load_migration_counts}
\title{Loads migration count data for SIMAH or CASCADE versions}
\usage{
load_migration_counts(SelectedState, DataDirectory)
}
\description{
Loads migration count data for SIMAH or CASCADE versions
}
\examples{
load_migration_counts
}
\keyword{counts}
\keyword{migration}
