% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_migration_rates.R
\name{load_migration_rates}
\alias{load_migration_rates}
\title{Loads migration rate data for SIMAH or CASCADE versions}
\usage{
load_migration_rates(SelectedState, WorkingDirectory)
}
\description{
Loads migration rate data for SIMAH or CASCADE versions
}
\examples{
load_migration_rates
}
\keyword{migration}
\keyword{rates}
