% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVACC.R
\name{MVACC}
\alias{MVACC}
\title{Calculates RR of motor vehicle injuries}
\usage{
MVACC(data, lhs)
}
\description{
Calculates RR of motor vehicle injuries
}
\examples{
MVACC mortality risk
}
\keyword{injuries}
\keyword{motor}
\keyword{vehicle}
