% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outward_migration.R
\name{outward_migration}
\alias{outward_migration}
\title{Deletes individuals due to migration in each simulated year}
\usage{
outward_migration(basepop, year_rates, y)
}
\description{
Deletes individuals due to migration in each simulated year
}
\examples{
outward_migration
}
\keyword{microsimulation}
