% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outward_migration_rate.R
\name{outward_migration_rate}
\alias{outward_migration_rate}
\title{Deletes individuals due to migration in each simulated year}
\usage{
outward_migration_rate(basepop, migration_rates, y)
}
\description{
Deletes individuals due to migration in each simulated year
}
\examples{
outward_migration
}
\keyword{microsimulation}
