% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_mortality.R
\name{postprocess_mortality}
\alias{postprocess_mortality}
\title{Calculates baseline mortality rates for different diseases}
\usage{
postprocess_mortality(DiseaseSummary, diseases, death_counts, inflation_factor)
}
\description{
Calculates baseline mortality rates for different diseases
}
\examples{
base rates
}
\keyword{base}
\keyword{calculate}
\keyword{rate}
