% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_alcohol_transition.R
\name{prob_alcohol_transition}
\alias{prob_alcohol_transition}
\title{Transitions alcohol consumption of individuals - through regression model}
\usage{
prob_alcohol_transition(data, model)
}
\description{
Transitions alcohol consumption of individuals - through regression model
}
\examples{
extract probabilities to transition from alcohol transition ordinal regression
}
\keyword{alcohol}
\keyword{microsimulation,}
