% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_brfss.R
\name{process_brfss}
\alias{process_brfss}
\title{Processes BRFSS data for new migrants and 18-year-olds entering the simulation}
\usage{
process_brfss(brfss, SelectedState)
}
\description{
Processes BRFSS data for new migrants and 18-year-olds entering the simulation
}
\examples{
process_brfss
}
\keyword{brfss}
