% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPF_functions.R, R/process_for_IPF.R
\name{process_for_IPF}
\alias{process_for_IPF}
\title{processes the BRFSS data for IPF for generating baseline population}
\usage{
process_for_IPF(data, selectedstate, variables)

process_for_IPF(data, selectedstate, variables)
}
\description{
processes the BRFSS data for IPF for generating baseline population

processes the BRFSS data for IPF for generating baseline population
}
\examples{
process_for_IPF
process_for_IPF
}
\keyword{brfss}
