% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_consumption.R
\name{reduce_consumption}
\alias{reduce_consumption}
\title{Reduce alcohol consumption for a given percentage reduction (to the whole population)}
\usage{
reduce_consumption(basepop, percentreduction)
}
\description{
Reduce alcohol consumption for a given percentage reduction (to the whole population)
}
\examples{
reduce_consumption
}
\keyword{alcohol}
\keyword{reduce}
