% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_individuals.R
\name{remove_individuals}
\alias{remove_individuals}
\title{Removes (uninflated) individuals from simulation}
\usage{
remove_individuals(data = basepop, disease, age_inflated, inflation_factors)
}
\description{
Removes (uninflated) individuals from simulation
}
\examples{
remove_individuals
}
\keyword{mortality}
\keyword{remove}
