% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microsim_alt_adjustmigration.R
\name{run_microsim_alt_adjustmigration}
\alias{run_microsim_alt_adjustmigration}
\title{Runs microsimulation - alternative version with different ordering
This version adjusts the migration COUNTS to match with the observed ACS data}
\usage{
run_microsim_alt_adjustmigration(
  seed,
  samplenum,
  basepop,
  brfss,
  death_counts,
  updatingeducation,
  education_transitions,
  migration_counts,
  population_counts,
  updatingalcohol,
  alcohol_transitions,
  catcontmodel,
  Hep,
  drinkingdistributions,
  base_counts,
  diseases,
  lhs,
  liverinteraction,
  policy = 0,
  percentreduction = 0.1,
  year_policy,
  inflation_factors,
  age_inflated,
  update_base_rate,
  minyear = 2000,
  maxyear = 2019,
  output = "demographics"
)
}
\description{
This function runs the microsimulation
}
\examples{
run_microsim_alt_adjustmigration
}
\keyword{microsimulation}
