% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microsim_determ.R
\name{run_microsim_determ}
\alias{run_microsim_determ}
\title{Runs microsimulation - deterministic version to get deterministic alcohol use pop}
\usage{
run_microsim_determ(
  seed,
  samplenum,
  basepop,
  brfss,
  death_counts,
  updatingeducation,
  education_transitions,
  migration_rates,
  updatingalcohol,
  alcohol_transitions,
  catcontmodel,
  drinkingdistributions,
  base_counts,
  diseases,
  lhs,
  sesinteraction,
  policy = 0,
  percentreduction = 0.1,
  year_policy,
  inflation_factors,
  age_inflated,
  update_base_rate,
  minyear = 2000,
  maxyear = 2005,
  output = "mortality"
)
}
\description{
This function runs the microsimulation
}
\examples{
run_microsim_determ
}
\keyword{microsimulation}
