% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_policy_parameters.R
\name{sample_policy_parameters}
\alias{sample_policy_parameters}
\title{Sample policy estimates}
\usage{
sample_policy_parameters(data, n_uncertainty)
}
\description{
This function simulates tax policy
}
\keyword{policy}
\keyword{tax}
