% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_base_counts.R
\name{setup_base_counts}
\alias{setup_base_counts}
\title{Sets up baseline mortality counts}
\usage{
setup_base_counts(death_counts, diseases, inflation_factors, age_categories)
}
\description{
This function sets up baseline mortality counts for COD modelled in simulation and inflates by a factor
}
\examples{
setup_base_counts
}
\keyword{baseline}
\keyword{mortality}
