% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_base_rates.R
\name{setup_base_rates}
\alias{setup_base_rates}
\title{Sets up baseline mortality rates}
\usage{
setup_base_rates(death_rates, diseases)
}
\description{
This function sets up baseline mortality rates for COD modelled in simulation
}
\examples{
setup_base_rates
}
\keyword{baseline}
\keyword{mortality}
