% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_education.R
\name{setup_education}
\alias{setup_education}
\title{Set up education data for simulation}
\usage{
setup_education(basepop, y)
}
\description{
This function sets up the education transition parameters for the simulation
}
\examples{
setup_education
}
\keyword{education}
