% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_education_covid.R
\name{setup_education_covid}
\alias{setup_education_covid}
\title{Setup education data for simulation}
\usage{
setup_education_covid(basepop, y)
}
\description{
This function sets up the education transition parameters for the simulation
}
\examples{
setup_education
}
\keyword{education}
