% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_for_IPF.R
\name{setup_for_IPF}
\alias{setup_for_IPF}
\title{gets the brfss data into the correct "individual" data format for IPF}
\usage{
setup_for_IPF(data, selectedstate, variables)
}
\description{
gets the brfss data into the correct "individual" data format for IPF
}
\examples{
setup_for_IPF
}
\keyword{brfss}
