% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mortality.R
\name{simulate_mortality}
\alias{simulate_mortality}
\title{Assigns risk for specific causes and simulates mortality}
\usage{
simulate_mortality(data, diseases)
}
\description{
This function simulates mortality
}
\examples{
simulate_mortality
}
\keyword{mortality}
