% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_alcohol_output.R
\name{summarise_alcohol_output}
\alias{summarise_alcohol_output}
\title{Compares alcohol consumption by subgroup to brfss data}
\usage{
summarise_alcohol_output(Output, SelectedState, WorkingDirectory)
}
\description{
Compares alcohol consumption by subgroup to brfss data
}
\examples{
summarise_alcohol_output
}
\keyword{alcohol}
\keyword{postprocessing}
