% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_alcohol_policy.R
\name{summarise_alcohol_policy}
\alias{summarise_alcohol_policy}
\title{Summarise mean alcohol consumption by policy setting}
\usage{
summarise_alcohol_policy(Output, SelectedState, version = standard)
}
\description{
Summarise mean alcohol consumption by policy setting
}
\keyword{alcohol}
\keyword{postprocessing}
