% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_education_output.R
\name{summarise_education_output}
\alias{summarise_education_output}
\title{Summarises microsimulation output by education and compares to target data}
\usage{
summarise_education_output(outputdata, SelectedState, WorkingDirectory)
}
\description{
This function computes summary statistics for educational attainment with optional plot
}
\examples{
summarise_education_output
}
\keyword{education}
