% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_hepatitis_output.R
\name{summarise_hepatitis_output}
\alias{summarise_hepatitis_output}
\title{Summarises microsimulation output for hepatitis B and C}
\usage{
summarise_hepatitis_output(outputdata)
}
\description{
This function computes summary statistics for hepatitis and plots the data
}
\examples{
summarise_hepatitis_output
}
\keyword{hepatitis}
