% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mortality_output.R
\name{summarise_mortality_output}
\alias{summarise_mortality_output}
\title{Compares mortality consumption by subgroup to brfss data}
\usage{
summarise_mortality_output(Output, SelectedState, WorkingDirectory, diseases)
}
\description{
Compares mortality consumption by subgroup to brfss data
}
\examples{
summarise_mortality_output
}
\keyword{mortality}
\keyword{postprocessing}
