% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mortality_output_agesp.R
\name{summarise_mortality_output_agesp}
\alias{summarise_mortality_output_agesp}
\title{Compares mortality consumption by subgroup to brfss data}
\usage{
summarise_mortality_output_agesp(
  Output,
  SelectedState,
  WorkingDirectory,
  inflation_factors,
  diseases
)
}
\description{
Compares mortality consumption by subgroup to brfss data
}
\examples{
summarise_mortality_output_agesp
}
\keyword{mortality}
\keyword{postprocessing}
