% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mortality_output_calibration.R
\name{summarise_mortality_output_calibration}
\alias{summarise_mortality_output_calibration}
\title{Compares mortality consumption by subgroup to brfss data}
\usage{
summarise_mortality_output_calibration(
  Output,
  SelectedState,
  inflation_factor,
  agestyear
)
}
\description{
Compares mortality consumption by subgroup to brfss data
}
\examples{
summarise_mortality_output
}
\keyword{mortality}
\keyword{postprocessing}
