% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mortality_output_policy.R
\name{summarise_mortality_output_policy}
\alias{summarise_mortality_output_policy}
\title{Compares mortality consumption by subgroup to brfss data}
\usage{
summarise_mortality_output_policy(
  Output,
  SelectedState,
  WorkingDirectory,
  inflation_factor,
  diseases
)
}
\description{
Compares mortality consumption by subgroup to brfss data
}
\examples{
summarise_mortality_output_policy
}
\keyword{mortality}
\keyword{postprocessing}
