% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_alcohol.R
\name{transition_alcohol}
\alias{transition_alcohol}
\title{Transitions alcohol consumption of individuals}
\usage{
transition_alcohol(data, transitions)
}
\description{
Transitions alcohol consumption of individuals
}
\examples{
transition_alcohol
}
\keyword{alcohol}
\keyword{microsimulation,}
