% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_alcohol_multinom_regression.R
\name{transition_alcohol_multinom_regression}
\alias{transition_alcohol_multinom_regression}
\title{Transitions alcohol consumption of individuals - through regression model}
\usage{
transition_alcohol_multinom_regression(data, model, y)
}
\description{
Transitions alcohol consumption of individuals - through regression model
}
\examples{
transition_alcohol_multinom_regression
}
\keyword{alcohol}
\keyword{microsimulation,}
