% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_ed.R
\name{transition_ed}
\alias{transition_ed}
\title{Transitions educational attainment of individuals}
\usage{
transition_ed(data, transitions)
}
\description{
Transitions educational attainment of individuals
}
\examples{
transition_ed
}
\keyword{education}
\keyword{microsimulation,}
