% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIJ.R
\name{UIJ}
\alias{UIJ}
\title{Calculates RR of other unintentional injuries}
\usage{
UIJ(data, lhs)
}
\description{
Calculates RR of other unintentional injuries
}
\examples{
UIJ mortality risk
}
\keyword{injuries}
\keyword{other}
\keyword{unintentional}
