% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_alcohol_cat.R
\name{update_alcohol_cat}
\alias{update_alcohol_cat}
\title{Updates alcohol category of individuals based on new gpd and new former drinker status}
\usage{
update_alcohol_cat(data)
}
\description{
Updates alcohol category of individuals based on new gpd and new former drinker status
}
\examples{
update_alcohol_cat
}
\keyword{alcohol}
\keyword{microsimulation,}
