% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_base_rate.R
\name{update_base_rate}
\alias{update_base_rate}
\title{Updates baseline mortality rates over time for different diseases according to parameter values}
\usage{
update_base_rate(rates, lhs, y)
}
\description{
Updates baseline mortality rates over time for different diseases according to parameter values
}
\examples{
updating base rates
}
\keyword{base}
\keyword{rate}
\keyword{update}
