% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_former_drinker.R
\name{update_former_drinker}
\alias{update_former_drinker}
\title{Updates new former drinker status}
\usage{
update_former_drinker(data)
}
\description{
Updates new former drinker status
}
\examples{
update_former_drinker
}
\keyword{alcohol}
\keyword{microsimulation,}
