% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating_alcohol.R
\name{updating_alcohol}
\alias{updating_alcohol}
\title{Updates alcohol use through matching with distributions from the BRFSS data}
\usage{
updating_alcohol(data, CDF, y, ...)
}
\description{
Updates alcohol use through matching with distributions from the BRFSS data
}
\examples{
update alcohol
}
\keyword{alcohol}
\keyword{update}
