% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microsim.R
\name{run_microsim}
\alias{run_microsim}
\title{Runs microsimulation}
\usage{
run_microsim(
  seed,
  samplenum,
  basepop,
  outwardmigrants,
  inwardmigrants,
  deathrates,
  apply_death_rates,
  updatingeducation,
  education_setup,
  transitionroles,
  calculate_migration_rates,
  outward_migration,
  inward_migration,
  brfss,
  Rates,
  AlctransitionProbability,
  transitions,
  PopPerYear,
  minyear,
  maxyear
)
}
\description{
This function allows you to express your love of cats.
}
\examples{
run_microsim
}
\keyword{microsimulation}
