% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microsim.R
\name{run_microsim}
\alias{run_microsim}
\title{Runs microsimulation}
\usage{
run_microsim(
  seed,
  samplenum,
  basepop,
  brfss,
  death_rates,
  updatingeducation,
  education_setup,
  migration_rates,
  updatingalcohol,
  alcohol_transitions,
  catcontmodel,
  policy = 0,
  percentreduction = 0.1,
  minyear = 2000,
  maxyear = 2019,
  output = "demographics"
)
}
\description{
This function runs the microsimulation
}
\examples{
run_microsim
}
\keyword{microsimulation}
