% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microsim_alt.R
\name{run_microsim_alt}
\alias{run_microsim_alt}
\title{Runs microsimulation - alternative version with different ordering}
\usage{
run_microsim_alt(
  seed,
  samplenum,
  basepop,
  brfss,
  death_counts,
  updatingeducation,
  education_transitions,
  COVID_specific_tps,
  migration_rates,
  updatingalcohol,
  alcohol_transitions,
  catcontmodel,
  drinkingdistributions,
  base_counts,
  diseases,
  lhs,
  sesinteraction,
  policy = 0,
  percentreduction = 0.1,
  year_policy,
  inflation_factors,
  age_inflated,
  update_base_rate,
  minyear,
  maxyear,
  output
)
}
\description{
This function runs the microsimulation
}
\examples{
run_microsim_alt
}
\keyword{microsimulation}
