% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_lhs.R
\name{sample_lhs}
\alias{sample_lhs}
\title{Samples parameters using a latin hypercube space filling design}
\usage{
sample_lhs(N_SAMPLES, PE)
}
\description{
This function generates a latin hypercube for a set number of parameters
}
\examples{
sample_lhs
}
\keyword{hypercube}
\keyword{latin}
