% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_alcohol_determ.R,
%   R/transition_alcohol_determ_alt.R
\name{transition_alcohol_determ}
\alias{transition_alcohol_determ}
\title{Transitions alcohol consumption of individuals - through deterministic method}
\usage{
transition_alcohol_determ(data, brfssdata, y)

transition_alcohol_determ(data, brfssdata, y)
}
\description{
Transitions alcohol consumption of individuals - through deterministic method

Transitions alcohol consumption of individuals - through deterministic method
}
\examples{
transition_alcohol_determ
transition_alcohol_determ
}
\keyword{alcohol}
\keyword{microsimulation,}
